pyproject.toml
[tool.poetry]
name = "obsidian-buttler"
version = "0.1.0"
description = ""
authors = ["Valter Nepomuceno <valter.nep@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
coloredlogs = "^15.0.1"
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
add_frontmatter = "obsidian_buttler.run:add_frontmatter_attribute_command"